Metro-Rail-Project-main/backend/requirements.txt
Django==4.2.20
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow<10.2
python-dotenv==1.0.0
qrcode==7.4.2
